Feature/sn trigger (#36)
* Increase SN pointing accuracy.

After the first SN direction estimation, the accuracy can be increased
by assuming a uniform distribution of event directions outside a cone
around the estmated direction. Only events inside the cone are actually
considered for the average direction calculation, and events outside the
cone are used to estimate the expectation value for the uniform
background inside the cone. That expectation value is then subtracted
from the average direction of all events in the cone.

* Minimal change to deal with no-float WCSim version.

Most single precision floats were eradicated and replaced with doubles
in WCSim/WCSim/#280. This is a minimal change to the TriggerInfo to
reflect that.